REF: Move cov_struct to simpler import.
