+++ statsmodels/genmod/cov_struct.py
