--- statsmodels/genmod/dependence_structures/__init__.py
@@ -1,4 +1,0 @@
-'''
-This module contains dependence structures for fitting models using
-generalized estimating equations (GEE).
-'''
