--- statsmodels/genmod/dependence_structures/api.py
@@ -1,2 +1,0 @@
-from covstruct import (Independence, Exchangeable, GlobalOddsRatio,
-                       Autoregressive, Nested, CovStruct)
