+++ statsmodels/genmod/tests/gee_gaussian_simulation_check.py
@@ -12,7 +12,7 @@ import numpy as np
 from itertools import product
 from statsmodels.genmod.families import Gaussian
 from statsmodels.genmod.generalized_estimating_equations import GEE
-from statsmodels.genmod.dependence_structures.api import Autoregressive, Nested
+from statsmodels.genmod.cov_struct import Autoregressive, Nested
 
 class GEE_simulator(object):
