+++ statsmodels/genmod/tests/gee_simulation_check.py
@@ -11,7 +11,7 @@ import numpy as np
 from itertools import product
 from statsmodels.genmod.families import Gaussian
 from statsmodels.genmod.generalized_estimating_equations import GEE
-from statsmodels.genmod.dependence_structures.api import Autoregressive, Nested
+from statsmodels.genmod.cov_struct import Autoregressive, Nested
 
 np.set_printoptions(formatter={'all': lambda x: "%8.3f" % x},
                     suppress=True)
